# api.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/DEVIL91053_1/coffee-pos/1.0.0
info:
  version: "1.0.0"
  title: Sample Password Flow OAuth2 Project
  description: >-
    This is an example of using OAuth2 Password Flow in a specification to
    describe security to your API.
security:
  - password:
      - read
      - write
paths:
...
  /menu/addItem:
    post:
      tags:
        - menu
      operationId: addItem
      description: add item to menu
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: successfully added item to menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /menu/addCategory:
    post:
      tags:
        - menu
      operationId: addCategory
      description: add product category
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: successfully added product category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /menu/discount:
    post:
      tags:
        - menu
      operationId: addDiscount
      description: add discount combination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Discount'
      responses:
        '200':
          description: successfully added new discount information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        status_code:
          type: integer
          format: int32
        message:
          type: string
          example: everything is ok
...
    Category:
      type: object
      properties:
        category_name:
          type: string
        menu_order:
          type: number
        description:
          type: string
    Item:
      type: object
      properties:
        product_name:
          type: string
          example: 'coffee'
        price:
          type: number
        category:
          type: string
        menu_order:
          type: number
        description:
          type: string
...
    Discount:
      type: object
      properties:
        discount_name:
          type: string
        combination:
          type: array
          items:
            $ref: '#/components/schemas/DiscountCombination'
        amount:
          type: number
    DiscountCombination:
      type: object
      properties:
        category_name:
          type: string
        quantity:
          type: number
...
  securitySchemes:
    password:
      type: oauth2
      flows:
        password:
          tokenUrl: 'http://example.com/oauth/token'
          scopes:
            write: allows modifying resources
            read: allows reading resources